Refactor to support implicit conversions to reduce verbosity in practical use
- Can now implicitly convert a Status to a Status<T>
- Can now implicitly convert a TResult to a Status<TResult>

This means that we no longer need to end methods with `return Status.AsSuccess(result);` we can simply `return result;` instead. This requires we have a return type of `Status<T>` instead of `IStatus<T>` with the same restrictions as `IActionResult` vs `ActionResult` from Asp.Net Core.

Similarly for returning an error status from a generic method, we no longer have to include the generic parameter as an implicit cast exists.